extract components from App
Extracted components for furniture and room data from App and into
separate files into components folder, respectively.
